feat：成功迁移存档（版本20260308->20260308.1

+++ Assets/Scripts/GeneralPreview/CDMV/DataBase.cs
@@ -26,7 +26,7 @@ public abstract class DataBase<TThis> : IDisposable
     public void Dispose() => RemoveAllCom();
     
     [ShowInInspector] readonly Dictionary<Type, ComBase> comDic = [];
-    float savedVersion = Const.Version;
+    double savedVersion = Const.Version;
     
     [DebuggerStepThrough]
     protected T AddCom<T>(T? com = null) where T : ComBase
